+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 ### New features
 * [#250](https://github.com/trema/pio/pull/250): Add Flow Removed message parser.
 * [#251](https://github.com/trema/pio/pull/251): Add Table Stats Request message generator.
+* [#266](https://github.com/trema/pio/pull/266): Add Port Stats Request message generator.
 
 ### Changes
 * [#265](https://github.com/trema/pio/pull/265): Rename `SetIpTos` -> `SetTos`.